int/disk/memory: organizes files.
